requirements-dev.txt
# AI Terminal Modular - Development Dependencies

# Core dependencies (from requirements.txt)
-r requirements.txt

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
coverage>=7.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
pre-commit>=3.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
myst-parser>=0.19.0

# Development Tools
ipython>=8.0.0
ipdb>=0.13.0
tox>=4.0.0

# Security
bandit>=1.7.0
safety>=2.3.0

# Performance
memory-profiler>=0.60.0
line-profiler>=4.0.0